'Exact' kerning
Kerning know computes a fixpoint to reach the exact distance.

The remaining ameliorations are
- take in account optical illusion : u_n o_2 do not look the same (very hard and for later)
- kerning between math and the text after or before (easy probably)

Moreover it is fast ...
- remain a problem : complexity quadratic in the depth of the formula because we recompute, resort too often
the list of bezier curves and bounding boxes for formulas. Formula are not very deep so we can live with that

The parameter precise_kerning in environment is of type float option.

None means : exact distance between Bezier curve
Some x means subdivise bezier curve until the thickness of the control polygone is below x.

x is by default 1e-2 and this makes no difference in the result (except for computing time).
None is really slow but is nice for debugging.