Configurable macros + others
New macros
\Configure_word_macro
\Configure_paragraph_macro
\Configure_environment

Allow to preprocess arguments and for word macro, to control the behavior with spaces.
See the end of DefaultGrammar.txp for examples.

\begin and \end allowed

\item do not need () anymore (thanks to \Configure_word_macro)

parskip with one newline + 2 spaces removed (it was really not a so good idea)
(kept 4 spaces at the beginning of a paragraph => center, but it shoud be replaced by tabulation indication, so it will
be removed too)

~ is now an unseccable white space

\{, \} \~ \\ can be used in text as escaped char