Macro arguments fixed
Macro with multiple arguments where broken !

It is now fixed.

To desambiguate, the solutions are

\m{arg1}(arg2)()(not an argument)
{\m{arg1}(arg2)}(not an argument) (math mode only)
\m{arg1}(arg2) (not an argument) (in math mode gives a product, in text mode requires "eat right" probably)
\m{arg1}(arg2}{not an argument} if the number of arguments was specified (see below)

New:

- The number of arguments can be specified with

\Configure_math_macros{ arity = 2 } (works for math, word, paragraph and environment)

Unfortunately, only desambiguate with curly brace because CAML arguments are parser as one token which is
preferred because of longuest match. Parsing CAML as a list of words should not be so hard and would solve the problem

- On can declare macro name

\Configure_math_macro{\bold}{ arity = 1 name = mathbold }

This allows the same macro name for one math macro and word macro (or one math macro and one paragraph macro).

Mouche:

- I protected "extras" arguments, this was wrong.