Click in pdf move in source !
This is controled by a flag --edit-link (off by default)

Add link with uri like "edit:filename@linenum@colnum" at many places 
in the code.

The precision is a group of words as define by the grammar
(words which go together in the same T) and whole math formulae.
not link in verbatim for instance.

More link can easily be added and precision can be imporved (to the price
of larger pdf).

Also added a python script xpdfclick and a sample .xpdfrc to use with xpdf.