Include ...
Include now works:

Compilation of document with includes will add the .cmx on the command line option
and consider the title (if any) as a relative section containing the whole document.

Author and institute are not used.

More than one level of includes is not tested ...

Patoline documentation is now Patoline.txp in the root folder.
if compiled with make doc and includes DefaultGrammar.cmx

DefaultGrammar.cmx is compiled by ocamlbuild, but ocamlbuild does not
computes the dependency, so it is removed at each make. 

mouche: --caml args where escaped in the wrong way.
still one may need to write:
    patoline --caml "-I \"folder with space\"" Patoline.txp