Drawing of glyphs
Drawing Bezier (caching a GlList for each glyph) is fast enough.
(Contrary to what one can read here:
http://stackoverflow.com/questions/5262951/what-is-state-of-the-art-for-text-rendering-in-opengl-as-of-version-4-1

But it renders the font too light ... Technics to solve this 
go to implicit curve (distance to the outline) and draw everything at
~0.5 pixels from the border.

So we should (in the future) precompute a distance matrix for each glyph
burk ... I will look for other solutions.