Fixing of bad size/base line of glyphs
Font have sometimes glyphs which are not compatible.
This happens when using together two math fonts, but even
inside AsanaMath, square bracket and white square bracket
do not have the same size not the same baseline.

A new function vkern_as in Box.ml can be used to adjust
the vertical kerning of one math symbol on another one.

See DefaultGrammar.txp.

The code for typesetting delimiters in now much simpler because 
it is not in charge of fixing bad glyphs anymore.