Better delimiters
- delimiter are now vertically centered, and a little upper (they where to low compared to
digits of capital letter). Better function to uniformise delimiter which are incompatible
between different fonts (and even inside asana !). Still some imperfection.

Tested on TD1 and my HDR. please report new problems.

Still the typography is not as good as it should be

- need smaller delimiter for lower case only formula like $cos(x)$

- need asymetric delimiter for un balanced fraction like $1 \over \int_a^2 f(t)dt$

- need more sizes because here we allow a 10% tolerance above and under delimiter in
display math to avoid to increase the size too fast.