Circular builds
This patch handles circular builds in the following way: if we try to
use a module that needs itself to be built, Patoline will just pretend
it was actually built. In the case of file item.txp using \item, this
is enough to avoid infinite loops and/or blocked mutexes.

Si Pierre peut trouver des exemples rigolos de graphes de dépendances…