Fix build of the Image driver
The Image driver depends on the GL driver. Yet, one cannot add
"Typography.GL" as a dependency of Image in the "needs" field in
configure.ml, since this would add a "-package Typography.GL" while
building Image. Yet this directive cannot work, and often causes a build
error, because the GL driver is not yet installed in a path where
ocamlfind can find it.

As a temporary hack, we list the GL driver as an internal dependency,
and add all external dependencies of GL to Image as well. A better fix
would recursively list external dependencies.