Diagrams: coord3d + Patonet: processes die after 5mn of inactivity (and reconnection of client works !)
Remark: if one reconnect to a modified patonet server by pressing a key, then main page and style are not reloaded and
the fonts are not present anymore. One should add something for the client to detect that the server has changed
while using it to force the client to reload the main page and style automatically.