change of semantics of nb_lines/err_lines in Interactive.ml
- no nb_lines/err_lines means the exact number of line of your code
- n > 0 means n (extra lines are hidden
- n < 0 means add n lines to your code

Beware that no check are done to detect if there is not enough room.

Remark about scrollbar: we might implement scrollbar, be we want full 
control of their look and therefore drawing a scroll bar in html is out of question.

More over, scroll bar should not have interaction with the server (sure ?) 
and therefore, the best is to create "clip_rectangle" with scroll bar
dranw as diagram.