Util out of Patoline and Typography. BREAKS PATOLINE - DO NOT PULL
This patch removes the files src/Typography/Util.ml/mli and src/Patoline/Util2.ml.
They are replaced by files in src/Util
Files in the Util directory are supposed to contain ocaml code that is likely to be
used in many components of patoline. They are our own extension to caml standard
library...
Everything compiles fine, however, this patck breaks patoline. This is a linking
problem that I am not able to solve. Maybe Christophe or Florian could have a look
at what I did in the Rule.mk files...