Fix environment problem in Maths.ml
The environment did not propagate to subscripts / superscripts, so the color
for instance, did not change according to the color of the "nucleus". I had
to remove a line that seemed quite obscure to me... I don't think I broke
anything but Christophe or PE should definitely review this patch. By the
way, I don't understand why we need to have a stack as an argument of the
function "draw".