pa_ocaml: bootstrap
pa_parser.ml: a rewriting of pa_glr.ml (a camlp4 extension for glr grammar) as a pa_ocaml extension.
pa_ocaml.ml is now able to compile itself.

Remains to test the result of this compilation and suppress the need for camlp4 (require some care,
not to loose the ability to compile pa_ocaml or pa_parser)

Remark: pa_parser.ml do not use quotation, because its initial version is not compiled with pa_ocaml
but with pa_glr. It could be good after bootstrap to simplify it with quotation.