Some work on the parsing or UnicodeData:
  - New library for permanent maps
  - unicode_parse renamed to pa_UnicodeData now uses the library to generate
    a database using the above library.
  - New file UCharInfo.ml is used to retrieve data from the library, it is
    part of unicodelib.

I have a linking problem on my machine with native code so I did not write the
Rules.mk part yet. There are instructions in the Rules.mk file if Christophe
wants to have a look...

For now, the generation of the database takes a lot of time (about 3hours
using bytecode) but a very naive approach is used for unicode ranges (i.e.
unicode characters that contain the same information) which are mostly
reserved part of the unicode range. I write an entry of the table for every
character, so we can win a factor 2 (maybe more) in both generation time and
file size.