Fixed the second problem.
The Net driver only needs to know where is the default plugin directory
(i.e. where the patonet.c file will be stored). As a workaround I created
a configuration file in rawlib. This is not a satisfactory fix, we really
design a better configuration system... I don't even know why it is working
right now (maybe it isn't). There are references with the same names in
Config.ml and in Config2.ml, an I cannot find any piece of code that links
the two. I think that most of the references in Config.ml are never taken
into account...