Remove -rectypes in Typography
The "-rectypes" option was necessary only to build
src/Typography/Break.ml type H.t. By adding a single constructor to this
type (which shouldn't change anything to the way values are stored in
memory), we can get rid of "-rectypes" everywhere in the code.

I confess that I do not understand the source code in Break.ml,
therefore the name "Hstruct.Val" of the constructor is absolutely not
explicit. Feel free to change it to something meaningful (the name does
not appear in the interface file, hence modifying it does not impact any
other part of Patoline).