Add missing dependencies in makefiles
Many .depends files were incomplete because the corresponding call to
ocamldep lacked the required "-I ..." include paths. This is a partial
fix, there are probably many other occurrences of this bug in other
makefiles. (We should ensure that ocamldep is always provided with the
same include paths as the ocamlc call which builds the corresponding
cmo/cmx files.)